+++ jest.config.ts
@@ -7,6 +7,6 @@ export default <Config>{
   moduleNameMapper: { '^~/(.*)$': '<rootDir>/$1' },
   rootDir: 'src',
   testEnvironment: 'node',
-  testRegex: '.*\\.spec\\.ts$',
+  testRegex: String.raw`.*\.spec\.ts$`,
   transform: { '^.+\\.(t|j)s$': 'ts-jest' },
 }