+++ src/user/user.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common'
+
+import { UserService } from '~/user/user.service'
+
+@Module({
+  providers: [UserService],
+})
+export class UserModule {}